Added signature to all built modules
We need to have signed modules. Otherwise kernel fails to load them. As
we are not using base kernel ( which does that implicitly on Bazel level
) - we need to add CONFIG_MODULE_SIG_ALL to our configuration.

This is some sort of workaround, as GKI build signs only a subset of
modules. Not all of them. But reaching that state without the usage of
the base kernel is an overhead. So we are using a simpler solution here.

Signed-off-by: Vladyslav Goncharuk <vladyslav_goncharuk@epam.com>